jwt integrated

+++ src/main/java/com/springproj/journalApp/service/journalentryservice.java
@@ -4,32 +4,41 @@ import com.springproj.journalApp.entity.journalentry;
 import com.springproj.journalApp.entity.user;
 import com.springproj.journalApp.repository.journalentryrepo;
 import org.bson.types.ObjectId;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.beans.Transient;
 import java.util.List;
 import java.util.Optional;
 
-@Component
+@Service
 //business logic ie. sari katha funtionalities yaha hogi
 public class journalentryservice {
 @Autowired
     private journalentryrepo  journalentryrepo;
 //this is a interface spring makes its implimentation during run time
 
+
+
+
 @Autowired
 private userservice userservice;
 @Transactional
     public void saveentry(journalentry je, String username){
+    try{
 user user = userservice.findByusername(username);
 //what if the user isnt saved ??-- the entry wil still be saved but not in any user we dont want this
         //so we will make the whole method a transaction
 
         journalentry saved = journalentryrepo.save(je);
         user.getJournalentries().add(saved);
-        userservice.saveentry(user);
+        userservice.saveuser(user);}
+    catch(Exception e){
+        throw new RuntimeException("an error occoured while saving the entry",e);
+
+    }
 
     }
 
@@ -40,11 +49,13 @@ user user = userservice.findByusername(username);
     public Optional<journalentry> findbyid(ObjectId id) {
         return journalentryrepo.findById(id);
     }
-
-    public void deltebyid(ObjectId id, String username){
-        user user = userservice.findByusername(username);
-       user.getJournalentries().removeIf(x->x.getId().equals(id));//this is lamda expression--java 8
-        //replacement of above
+@Transactional
+    public boolean deltebyid(ObjectId id, String username){
+     boolean removed= false;
+      try{
+          user user = userservice.findByusername(username);
+           removed=  user.getJournalentries().removeIf(x->x.getId().equals(id));//this is lamda expression--java 8
+          //replacement of above
 //        Iterator<journalentry> iterator = user.getJournalentries().iterator();
 //
 //        while (iterator.hasNext()) {
@@ -53,9 +64,20 @@ user user = userservice.findByusername(username);
 //                iterator.remove();
 //            }
 //        }
+          if(removed){
+              userservice.saveuser(user);
+              journalentryrepo.deleteById(id);
+          }
 
-userservice.saveentry(user);
-        journalentryrepo.deleteById(id);
+
+      }
+      catch (Exception e){
+          System.out.println(e);
+          throw new RuntimeException("a error occured while deleting the entry");
+      }
+    return removed;
+
+
 }
 
 
@@ -66,6 +88,7 @@ public void saveentry(journalentry je){
 journalentryrepo.save(je);
 
 }
+
